Fix response code in Controllers, fix longValue in calculator.
    - Fixed response code for positive result (from FOUND[302] to OK [200], since 302 is used for redirecting to resource),
    - Fixed function which should return decimal value instead of long (integer value).

+++ backend/stock-market/src/main/java/pl/kazet/stockmarket/utils/NetProfitMarginCalculator.java
@@ -56,7 +56,7 @@ public class NetProfitMarginCalculator {
     }
 
     private  BigDecimal getTotalCurrentSelloutValue(){
-        return BigDecimal.valueOf(amountOfBoughtStocks * currentStockPrice.getCurrentSelloutPrice().longValue());
+        return BigDecimal.valueOf(amountOfBoughtStocks * currentStockPrice.getCurrentSelloutPrice().doubleValue());
     }
 
     private BigDecimal getSumProductOfBoughtStocks(){
